Refactor info-json serving classes, set Id property per request
