+++ Orchestrator.Tests/Settings/OrchestratorSettingsTests.cs
@@ -39,38 +39,6 @@ namespace Orchestrator.Tests.Settings
                 }
             };
         }
-        
-        [Theory]
-        [InlineData("2.1", IIIF.Presentation.Version.V2)]
-        [InlineData("3.0", IIIF.Presentation.Version.V3)]
-        [InlineData("something else", IIIF.Presentation.Version.V3)]
-        public void GetDefaultIIIFPresentationVersion_Correct(string value, IIIF.Presentation.Version expected)
-        {
-            // Arrange
-            var settings = new OrchestratorSettings { DefaultIIIFPresentationVersion = value };
-            
-            // Act
-            var actual = settings.GetDefaultIIIFPresentationVersion();
-            
-            // Assert
-            actual.Should().Be(expected);
-        }
-        
-        [Theory]
-        [InlineData("2.1", IIIF.ImageApi.Version.V2)]
-        [InlineData("3.0", IIIF.ImageApi.Version.V3)]
-        [InlineData("something else", IIIF.ImageApi.Version.V3)]
-        public void GetDefaultIIIFImageVersion_Correct(string value, IIIF.ImageApi.Version expected)
-        {
-            // Arrange
-            var settings = new OrchestratorSettings { DefaultIIIFImageVersion = value };
-            
-            // Act
-            var actual = settings.GetDefaultIIIFImageVersion();
-            
-            // Assert
-            actual.Should().Be(expected);
-        }
 
         [Theory]
         [InlineData(ImageServer.Cantaloupe, "/iiif/3/%2F99%2F100%2Fte%2Fst%2F-i%2Fma%2Ftest-image-name%2Ftest-image-name")]
